fix setup user profile
