+++ frontend/src/api/health.ts
@@ -1,4 +1,4 @@
 export async function getHealth() {
-  const res = await fetch(`/health`);
+  const res = await fetch(`/api/health`);
   return res.json();
 }
